+++ release_notes.md
@@ -21,6 +21,7 @@ PREESM Changelog
 * Refactor: TwinBuffer now is called DistributedBuffer;
 * Hardware Codegen Refactor: CHardwarePrinter extends now the CPrinter and override some of the methods;
 * SPiDER codegen considers empty PAPIFY configs as papify=false;
+* Update round buffer memory script to handle multiple inputs;
 
 ### Bug fix
 * fix #186